Querying our index
The default language model is OpenAI, so first we load a local one: [Phi3 from Ollama](https://github.com/ollama/ollama). Ollama is a library that enables us to run small language models locally. We set this in the global `Settings` object.

Once our LLM is loaded, we can create a `QueryEngine`. The `QueryEngine` ties all the components together: we can query it, then the embedding model creates an embedding vector of the query, and with that vector, we do a similarity search on the embeddings created from chunks stored in the index. Then the llm creates our response using the most relevant chunks.