Persisting storage
With previous piece of code, the index is created every time our code is executed. This is inefficient. We only need to create the chunks and the embeddings once, then we can query them as much as we want.

So here with if clause, we check if our `PERSIST_DIR` already exists, and if not, we persist our data in it. And next time, we load our index from that directory.