Changing `top_k`
The `top_k` parameter controls how many chunks will be passed to the LLM to answer our query. TO modify this paramter, we need to create a `VectorIndexRetriever` object in LlamaIndex before creating our `QueryEngine`. On this object, we can set the `similarity_top_k` parameter. The default is 1, we set it to 3 here.

After answering our query, we print our the chunks used to answer our query. This is in the `source_nodes` attribute of our response object.