Ingestion pipeline moved into other file
We refactored our code a bit. Now, we have an `ingestion.py` file, where the outcome is to create an index, persisted in the storage folder. In this file, we first create a log file, then set the persist directory in the `PERSIST_DIR` constant. Then we need to define our embedding model, as it is needed to create the embedding models.

We configure chunking hyperparameters, create our index from the files in the data folder, and persist it.

At the beginning of the file, we created a separate folder for logs.

In the `llamaindex_demo.py` file, we removed the code that is needed to create our index, we only load it. This might be better to do in a try-except block, but now we assume our `PERSIST_DIR` exists.