Created a web application using our RAG
Now we used the chainlit python library to create a simple web app. Chainlit is built on top of streamlit, exclusively to build LLM applications with it. [Read more about chainlit in the docs](https://docs.chainlit.io/get-started/overview).

The `start` function is called when the web app is entered. So here, we set what embedding and language model we are using, load our index and create a query engine from it.

The `main` function, due to the `on_message` decorator, is called only when the user enters a prompt. It calls the query engine to generate a response, and returns this response to the user.