Removed `.env` file, created `config.toml`
During development, it's better practice to create a config file, that creates publicly accessible informations, such as RAG hyperparams like what embedding model or language model to use. TOML is a config file format(Tom's Obvious Minimal Language) that is commonly used with Python. The `toml` Python module is required to read such files. It is easy to read, as you can define section names, that can be related to different parts of the RAG pipeline, for example, chunking.

`.env` files can be used to store sensitie information, such as secrets, like the OpenAI API key. This file should not be kept in the repo. A better alternative is to use a secret manager.