Using QDrant VectorDB as storage context
So far, we have used local disk persistance for our embedding vectors.

Now, we use QDrant vector database. QDrant runs on a kubernetes cluster in Azure in our setup. We need to set its URL, port and the collection name we want to create. These are in the `config.toml` file. A vector database creates an effient index, that can quickly find chunks with the most semantic simiarity. However, this effect is only relevant when we have chunks in the range of millions.